Remove redundant files
Since we are migrating the original lightbox plugin over to bootstrap there are a number of files that are no longer needed. This includes the css for the example page, the releases directory (and the ds_store file that somehow got added along with it), and the jquery files (we'll be using the cdn versions instead).