Include images as base64 data URIs
By including the images as data URIs we can reduce the number of http requests and no longer have to concerned as to which directory the image files used by this plugin are located. This does mean however, that IE7 and lower are no longer supported directly. If support for these browsers are required I would recomend checking this page out http://www.phpied.com/mhtml-when-you-need-data-uris-in-ie7-and-under/ , alternatively the images can be re-added from the original lightbox project http://github.com/lokesh/lightbox2/