Add support for title links
Fix Issue #2. Updated the js files so that titles can now act as links via the data-title-link attribute. By default the link opens a new window with the exception of hash links, they open on the same page without reloading. I've also updated the readme and example pages to show you how to implement it, updated the download link and fixed a spelling misake.